dbScrapperV3.1
The dbScrapperV3.1 is a script used as a module in a python based program.
It handles requests to an API and parses the data to be able to insert it into a MySQL database server.
The script is able to query data from the database based on a _uniqueId_ specified in its configuration file, comparing if the data is the same. It can update or create new entries within the database depending on the comparisson result between the data.